+++ src/common/iServer/BufferAnalystService.js
@@ -87,26 +87,6 @@ export class BufferAnalystService extends SpatialAnalystBase {
         });
     }
 
-
-    /**
-     * @method SuperMap.BufferAnalystService.prototype.toGeoJSONResult
-     * @description 将含有geometry的数据转换为geojson格式。
-     * @param result - {Object} 服务器返回的结果对象。
-     */
-    // toGeoJSONResult(result) {
-    //     if (!result) {
-    //         return result;
-    //     }
-    //
-    //     var analystResult = super.toGeoJSONResult(result);
-    //     if (analystResult.resultGeometry) {
-    //         var geoJSONFormat = new GeoJSON();
-    //         result = JSON.parse(geoJSONFormat.write(analystResult.resultGeometry));
-    //     }
-    //     return result;
-    // }
-
-
 }
 
 SuperMap.BufferAnalystService = BufferAnalystService;