+++ src/openlayers/mapping/SuperMapCloud.js
@@ -19,17 +19,32 @@ export class SuperMapCloud extends ol.source.XYZ {
         var mapType = options.mapType || 'web';
         var url = options.url || 'http://t2.supermapcloud.com/FileService/image?map={mapName}&type={type}&x={x}&y={y}&z={z}';
         url = url.replace('{mapName}', mapName).replace('{type}', mapType);
-        super({
+
+        var superOptions = {
             attributions: attributions,
             cacheSize: options.cacheSize,
             crossOrigin: options.crossOrigin,
             opaque: options.opaque === undefined ? true : options.opaque,
             maxZoom: options.maxZoom || 18,
             reprojectionErrorThreshold: options.reprojectionErrorThreshold,
-            tileLoadFunction: options.tileLoadFunction,
             url: url,
             wrapX: options.wrapX
-        });
+        };
+        //需要代理时走自定义 tileLoadFunction，否则走默认的tileLoadFunction
+        if (options.tileProxy) {
+            superOptions.tileLoadFunction = tileLoadFunction;
+        }
+        super(superOptions);
+
+        if (options.tileProxy) {
+            this.tileProxy = options.tileProxy;
+        }
+        //需要代理时，走以下代码
+        var me = this;
+        function tileLoadFunction(imageTile, src) {
+            //支持代理
+            imageTile.getImage().src = me.tileProxy + encodeURIComponent(src);
+        }
     }
 }